requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
dateutils==0.6.6
Django==2.1.1
django-filter==2.0.0
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
kombu==4.2.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
vine==1.1.4
